Updates build workflow and adds mise config
Updates GitHub Actions build workflow to use specific versions of actions for increased reliability and reproducibility.

Adds a `mise.toml` file to configure Node.js and pnpm versions.

Removes `coverage/lcov.info` from the repository.  This file is generated during the build process and does not need to be tracked.